Made the game easy mode

+++ Tempo/constants.h
@@ -6,8 +6,8 @@ const int COLOR_MODE_MULTI = 1;
 const int SCREEN_WIDTH = 1280;
 const int SCREEN_HEIGHT = 800;
 const int SCREEN_BPP = 32;
-const int Z_FAR = 100000;
-const int Z_NEAR = 1000;
+const float Z_FAR = 100000.0f;
+const float Z_NEAR = 1000.f;
 const float FPS_CAP = 1000.f / 60;
 
 const float SHIFT_INTERVAL = 10.f;
